+++ pycaret/internal/pycaret_experiment/time_series_experiment.py
@@ -1,12 +1,7 @@
-from pandas.core.frame import DataFrame
 from sktime.forecasting.model_selection import (
     ExpandingWindowSplitter,
     SlidingWindowSplitter,
 )
-
-import statsmodels.api as sm
-from statsmodels.tsa.seasonal import seasonal_decompose, STL
-from statsmodels.tsa.stattools import acf, pacf
 
 
 from pycaret.internal.pycaret_experiment.utils import highlight_setup, MLUsecase
@@ -57,7 +52,8 @@ from joblib import Parallel, delayed  # type: ignore
 from sktime.utils.validation.forecasting import check_y_X  # type: ignore
 from sktime.forecasting.model_selection import SlidingWindowSplitter  # type: ignore
 
-import pmdarima as pm
+from ..tests.time_series import test_
+
 
 warnings.filterwarnings("ignore")
 LOGGER = get_logger()
@@ -3291,342 +3287,10 @@ class TimeSeriesExperiment(_SupervisedExperiment):
 
     def _get_y_data(self, split="all"):
         if split == "all":
-            data = self.y.copy()
+            data = self.y
         elif split == "train":
-            data = self.y_train.copy()
+            data = self.y_train
         else:
             raise ValueError("split value: '{split}' is not supported.")
         return data
-
-
-#############################################################
-###################### TESTS STARTS HERE ####################
-#############################################################
-
-#################
-#### Helpers ####
-#################
-
-
-def test_(data, test: str, alpha: float = 0.05, *kwargs):
-    if test == "all":
-        results = test_all(data=data, alpha=alpha)
-    elif test == "stat_summary":
-        results = test_summary_statistics(data=data)
-    elif test == "white_noise":
-        results = test_is_white_noise(data=data, alpha=alpha, verbose=True)[1]
-    elif test == "stationarity":
-        results = test_is_stationary(data=data, alpha=alpha)
-    elif test == "adf":
-        results = test_is_stationary_adf(data=data, alpha=alpha, verbose=True)[1]
-    elif test == "kpss":
-        results = test_is_stationary_kpss(data=data, alpha=alpha, verbose=True)[1]
-    elif test == "normality":
-        results = test_is_gaussian(data=data, alpha=alpha, verbose=True)[1]
-    else:
-        raise ValueError(f"Tests: '{test}' is not supported.")
-    return results
-
-
-########################
-#### Combined Tests ####
-########################
-def test_all(data, alpha: float = 0.05):
-    result_summary_stats = test_summary_statistics(data=data)
-    result_summary_stats["Setting"] = ""
-    result_summary_stats.set_index("Setting", append=True, inplace=True)
-
-    result_wn = test_is_white_noise(data, verbose=True)
-    result_adf = test_is_stationary_adf(data, alpha=alpha, verbose=True)
-    result_kpss = test_is_stationary_kpss(data, alpha=alpha, verbose=True)
-    result_normality = test_is_gaussian(data, alpha=alpha, verbose=True)
-
-    all_dfs = [
-        result_summary_stats,
-        result_wn[1],
-        result_adf[1],
-        result_kpss[1],
-        result_normality[1],
-    ]
-    final = pd.concat(all_dfs)
-    return final
-
-
-def test_is_stationary(data, alpha: float = 0.05):
-    result_adf = test_is_stationary_adf(data, alpha=alpha, verbose=True)
-    result_kpss = test_is_stationary_kpss(data, alpha=alpha, verbose=True)
-
-    all_dfs = [
-        result_adf[1],
-        result_kpss[1],
-    ]
-    final = pd.concat(all_dfs)
-    return final
-
-
-##########################
-#### Individual Tests ####
-##########################
-
-
-def test_summary_statistics(data: pd.Series):
-    distinct_counts = dict(data.value_counts(normalize=True))
-    result = {
-        "Length": len(data),
-        "Mean": data.mean(),
-        "Median": data.median(),
-        "Standard Deviation": data.std(),
-        "Variance": data.var(),
-        "Kurtosis": data.kurt(),
-        "Skewness": data.skew(),
-        "# Distinct Values": len(distinct_counts),
-    }
-
-    result = pd.DataFrame(result, index=["Value"]).T.reset_index()
-    result = _format_test_results(result, "Summary", "Statistics")
-
-    return result
-
-
-def test_is_white_noise(
-    data: pd.Series,
-    lags: List[int] = [24, 48],
-    alpha: float = 0.05,
-    verbose: bool = False,
-):
-    """Performs the Ljung-Box test for testing if a time series is White Noise
-
-    H0: The data is consistent with white noise
-    Ha: The data is not consistent with white noise.
-
-    Parameters
-    ----------
-    data : pd.Series
-        The time series that has to be tested
-    lags : List[int], optional
-        The lags used to test the autocorelation for white noise, by default [24, 48]
-    """
-    results = sm.stats.acorr_ljungbox(data, lags=lags, return_df=True)
-    results["White Noise"] = results["lb_pvalue"] > alpha
-    is_white_noise = False if results["White Noise"].all() == False else True
-    results.rename(
-        columns={"lb_stat": "Test Statictic", "lb_pvalue": "p-value"}, inplace=True,
-    )
-
-    results.reset_index(inplace=True)
-    results.rename(columns={"index": "Setting"}, inplace=True)
-    # TODO: Add alpha value to Settings
-    results = pd.melt(results, id_vars="Setting", var_name="index", value_name="Value")
-    results = _format_test_results(results, "White Noise", "Ljung-Box")
-
-    if verbose:
-        return is_white_noise, results
-    else:
-        return is_white_noise
-
-
-def test_is_stationary_adf(data: pd.Series, alpha: float = 0.05, verbose: bool = False):
-    """Checks Difference Stationarity
-
-    H0: The time series is not stationary (has a unit root)
-    H1: The time series is stationary
-
-    If not stationary, try differencing
-
-    Parameters
-    ----------
-    data : pd.Series
-        Time Series to be tested
-    """
-    from statsmodels.tsa.stattools import adfuller
-
-    results = adfuller(data, autolag="AIC", maxlag=20)
-    test_statistic = results[0]
-    critical_values = results[4]
-    p_value = results[1]
-    stationarity = True if p_value < alpha else False
-    details = {
-        "Stationarity": stationarity,
-        "p-value": p_value,
-        "Test Statistic": test_statistic,
-    }
-    critical_values = {
-        f"Critical Value {key}": value for key, value in critical_values.items()
-    }
-    details.update(critical_values)
-
-    details = pd.DataFrame(details, index=["Value"]).T.reset_index()
-    details["Setting"] = alpha
-    details = _format_test_results(details, "Stationarity", "ADF")
-
-    if verbose:
-        return stationarity, details
-    else:
-        return stationarity
-
-
-def test_is_stationary_kpss(
-    data: pd.Series, alpha: float = 0.05, verbose: bool = False
-):
-    """Checks Stationarity around a deterministic trend
-
-    H0: The time series is trend stationary
-    H1: The time series is not trend stationary
-
-    If not trend stationary, then try removing trend.
-
-    Parameters
-    ----------
-    data : pd.Series
-        Time Series to be tested
-    """
-    from statsmodels.tsa.api import kpss
-
-    results = kpss(data, regression="ct", nlags="auto")
-    test_statistic = results[0]
-    p_value = results[1]
-    critical_values = results[3]
-    stationarity = False if p_value < alpha else True
-    details = {
-        "Trend Stationarity": stationarity,
-        "p-value": p_value,
-        "Test Statistic": test_statistic,
-    }
-    critical_values = {
-        f"Critical Value {key}": value for key, value in critical_values.items()
-    }
-    details.update(critical_values)
-
-    details = pd.DataFrame(details, index=["Value"]).T.reset_index()
-    details["Setting"] = alpha
-    details = _format_test_results(details, "Stationarity", "KPSS")
-
-    if verbose:
-        return stationarity, details
-    else:
-        return stationarity
-
-
-def test_trend():
-    pass
-
-
-def test_seasonality():
-    pass
-
-
-def test_is_gaussian(data: pd.Series, alpha: float = 0.05, verbose: bool = False):
-    """Performs the shapiro test to check for normality of data
-    """
-    from scipy.stats import shapiro, norm
-
-    p_value = shapiro(data.values.squeeze())[1]
-    normality = True if p_value > alpha else False
-    details = {
-        "Normality": normality,
-        "p-value": p_value,
-    }
-    details = pd.DataFrame(details, index=["Value"]).T.reset_index()
-    details["Setting"] = alpha
-    details = _format_test_results(details, "Normality", "Shapiro")
-
-    if verbose:
-        return normality, details
-    else:
-        return normality
-
-
-def recommend_d(data: pd.Series, **kwargs) -> int:
-    """Returns the recommended value of differencing order 'd' to use
-
-    Parameters
-    ----------
-    data : pd.Series
-        The data for which the differencing order needs to be calculated
-
-    *kwargs: Keyword arguments that can be passed to the difference test.
-        Values are:
-            alpha : float, optional
-                Significance Value, by default 0.05
-            test : str, optional
-                The test to use to test the order of differencing, by default 'kpss'
-            max_d : int, optional
-                maximum differencing order to try, by default 2
-
-    Returns
-    -------
-    int
-        The differencing order to use
-    """
-    recommended_d = pm.arima.ndiffs(data, **kwargs)
-    return recommended_d
-
-
-def recommend_D(data: pd.Series, sp: int, **kwargs) -> int:
-    """Returns the recommended value of differencing order 'd' to use
-
-    Parameters
-    ----------
-    data : pd.Series
-        The data for which the differencing order needs to be calculated
-
-    sp : int
-        The number of seasonal periods (i.e., frequency of the time series)
-
-    *kwargs: Keyword arguments that can be passed to the difference test.
-        Values are:
-            alpha : float, optional
-                Significance Value, by default 0.05
-            test : str, optional
-                Type of unit root test of seasonality to use in order to detect
-                seasonal periodicity. Valid tests include (“ocsb”, “ch”).
-                Note that the CHTest is very slow for large data.
-            max_D : int, optional
-                maximum seasonal differencing order to try, by default 2
-
-    Returns
-    -------
-    int
-        The differencing order to use
-    """
-    recommended_D = pm.arima.ndsiffs(data, m=sp, **kwargs)
-    return recommended_D
-
-
-def _format_test_results(
-    result: pd.DataFrame, test: str, test_name: str
-) -> pd.DataFrame:
-    """Formats the results of a test into a multiindex dataframe with first
-    level = "test" and second level = "name". The final dataframe has a column
-    called value corresponding to the "name"
-
-    Parameters
-    ----------
-    result : pd.DataFrame
-        The unformatted result dataframe that has to be formatted.
-        Must contain
-          - a column called "index" corresponding to the
-            values (test statistic) to be used for the second level.
-          - a column called "value" which contains the
-            value of the test statistic.
-        Optionally
-          - a column called "Setting" which can contain information
-            such as alpha value used, or number of lags used
-
-    test_name : str
-        The name of the test. Used as the value for the first level
-
-    Returns
-    -------
-    pd.DataFrame
-        [description]
-    """
-    result.rename(columns={"index": "Property"}, inplace=True)
-    result["Test"] = test
-    result["Test Name"] = test_name
-    if "Setting" in result.columns:
-        result.set_index(["Test", "Test Name", "Property", "Setting"], inplace=True)
-    else:
-        result.set_index(["Test", "Test Name", "Property"], inplace=True)
-    return result
